refactor sidebars and decouple garment reset
